Move left over conditions funcs into their own file
